moved initial data function to its own folder
